refactor(governance): 拆分 runtime ledger 持久化模块
