+++ src-tauri/src/runtime/ledger.rs
@@ -1,0 +1,51 @@
+use std::fs;
+use std::io::Write;
+use std::path::Path;
+
+use serde::{Deserialize, Serialize};
+
+use super::pool_types::{RuntimePoolDiagnostics, RuntimePoolRow};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub(super) struct PersistedRuntimeLedger {
+    pub(super) rows: Vec<RuntimePoolRow>,
+    pub(super) diagnostics: RuntimePoolDiagnostics,
+}
+
+pub(super) fn write_json_atomically(path: &Path, content: &str) -> Result<(), String> {
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent).map_err(|error| error.to_string())?;
+    }
+    let parent = path
+        .parent()
+        .ok_or_else(|| format!("runtime ledger path has no parent: {}", path.display()))?;
+    let filename = path
+        .file_name()
+        .and_then(|value| value.to_str())
+        .ok_or_else(|| {
+            format!(
+                "runtime ledger path has invalid filename: {}",
+                path.display()
+            )
+        })?;
+    let temp_path = parent.join(format!(".{filename}.{}.tmp", uuid::Uuid::new_v4()));
+    let mut temp_file = fs::OpenOptions::new()
+        .create_new(true)
+        .write(true)
+        .open(&temp_path)
+        .map_err(|error| error.to_string())?;
+    temp_file
+        .write_all(content.as_bytes())
+        .map_err(|error| error.to_string())?;
+    temp_file.sync_all().map_err(|error| error.to_string())?;
+    #[cfg(target_os = "windows")]
+    if path.exists() {
+        fs::remove_file(path).map_err(|error| error.to_string())?;
+    }
+    if let Err(error) = fs::rename(&temp_path, path) {
+        let _ = fs::remove_file(&temp_path);
+        return Err(error.to_string());
+    }
+    Ok(())
+}
